+++ objects/src/assets/fungible.rs
@@ -1,11 +1,14 @@
 use alloc::string::ToString;
 use core::fmt;
 
-use vm_core::FieldElement;
+use vm_core::{
+    utils::{ByteReader, ByteWriter, Deserializable, Serializable},
+    FieldElement,
+};
+use vm_processor::DeserializationError;
 
 use super::{
-    is_not_a_non_fungible_asset, parse_word, AccountId, AccountType, Asset, AssetError, Felt, Word,
-    ZERO,
+    is_not_a_non_fungible_asset, AccountId, AccountType, Asset, AssetError, Felt, Word, ZERO,
 };
 
 // FUNGIBLE ASSET
@@ -145,16 +148,6 @@ impl From<FungibleAsset> for Word {
     }
 }
 
-impl From<FungibleAsset> for [u8; 32] {
-    fn from(asset: FungibleAsset) -> Self {
-        let mut result = [0_u8; 32];
-        let id_bytes: [u8; 8] = asset.faucet_id.into();
-        result[..8].copy_from_slice(&asset.amount.to_le_bytes());
-        result[24..].copy_from_slice(&id_bytes);
-        result
-    }
-}
-
 impl From<FungibleAsset> for Asset {
     fn from(asset: FungibleAsset) -> Self {
         Asset::Fungible(asset)
@@ -175,17 +168,83 @@ impl TryFrom<Word> for FungibleAsset {
     }
 }
 
-impl TryFrom<[u8; 32]> for FungibleAsset {
-    type Error = AssetError;
-
-    fn try_from(value: [u8; 32]) -> Result<Self, Self::Error> {
-        let word = parse_word(value)?;
-        Self::try_from(word)
-    }
-}
-
 impl fmt::Display for FungibleAsset {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{:?}", self)
+    }
+}
+
+// SERIALIZATION
+// ================================================================================================
+
+impl Serializable for FungibleAsset {
+    fn write_into<W: ByteWriter>(&self, target: &mut W) {
+        // All assets should serialize their faucet ID at the first position to allow them to be
+        // easily distinguishable during deserialization.
+        target.write(self.faucet_id);
+        target.write(self.amount);
+    }
+
+    fn get_size_hint(&self) -> usize {
+        self.faucet_id.get_size_hint() + self.amount.get_size_hint()
+    }
+}
+
+impl Deserializable for FungibleAsset {
+    fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
+        let faucet_id: AccountId = source.read()?;
+        FungibleAsset::deserialize_with_account_id(faucet_id, source)
+    }
+}
+
+impl FungibleAsset {
+    /// Deserializes a [`FungibleAsset`] from an [`AccountId`] and the remaining data from the given
+    /// `source`.
+    pub(super) fn deserialize_with_account_id<R: ByteReader>(
+        faucet_id: AccountId,
+        source: &mut R,
+    ) -> Result<Self, DeserializationError> {
+        let amount: u64 = source.read()?;
+        FungibleAsset::new(faucet_id, amount)
+            .map_err(|err| DeserializationError::InvalidValue(err.to_string()))
+    }
+}
+
+// TESTS
+// ================================================================================================
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::accounts::account_id::testing::{
+        ACCOUNT_ID_FUNGIBLE_FAUCET_OFF_CHAIN, ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN,
+        ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_1, ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_2,
+        ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_3, ACCOUNT_ID_NON_FUNGIBLE_FAUCET_OFF_CHAIN,
+    };
+
+    #[test]
+    fn test_fungible_asset_serde() {
+        for fungible_account_id in [
+            ACCOUNT_ID_FUNGIBLE_FAUCET_OFF_CHAIN,
+            ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN,
+            ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_1,
+            ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_2,
+            ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_3,
+        ] {
+            let account_id = AccountId::try_from(fungible_account_id).unwrap();
+            let fungible_asset = FungibleAsset::new(account_id, 10).unwrap();
+            assert_eq!(
+                fungible_asset,
+                FungibleAsset::read_from_bytes(&fungible_asset.to_bytes()).unwrap()
+            );
+        }
+
+        let account_id = AccountId::try_from(ACCOUNT_ID_FUNGIBLE_FAUCET_ON_CHAIN_3).unwrap();
+        let asset = FungibleAsset::new(account_id, 50).unwrap();
+        let mut asset_bytes = asset.to_bytes();
+        // Set invalid Faucet ID.
+        asset_bytes[0..8].copy_from_slice(&ACCOUNT_ID_NON_FUNGIBLE_FAUCET_OFF_CHAIN.to_le_bytes());
+        let err = FungibleAsset::read_from_bytes(&asset_bytes).unwrap_err();
+        assert!(matches!(err, DeserializationError::InvalidValue(_)));
     }
 }