+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - [BREAKING] Moved `MAX_NUM_FOREIGN_ACCOUNTS` into `miden-objects` (#904).
 - Implemented `storage_size`, updated storage bounds (#886).
 - [BREAKING] Auto-generate `KERNEL_ERRORS` list from the transaction kernel's MASM files and rework error constant names (#906).
+- Implement `Serializable` for `FungibleAsset` (#907).
 
 ## 0.5.1 (2024-08-28) - `miden-objects` crate only